add mobile responsiveness
