+++ src/interfaces/assistants_web/src/app/(main)/(chat)/Chat.tsx
@@ -1,10 +1,8 @@
 'use client';
 
-import { Transition } from '@headlessui/react';
 import { useEffect } from 'react';
 
 import { Document, ManagedTool } from '@/cohere-client';
-import AgentRightPanel from '@/components/Agents/AgentRightPanel';
 import { ConnectDataModal } from '@/components/ConnectDataModal';
 import Conversation from '@/components/Conversation';
 import { ConversationError } from '@/components/ConversationError';
@@ -12,12 +10,11 @@ import { Spinner } from '@/components/Shared';
 import { TOOL_PYTHON_INTERPRETER_ID } from '@/constants';
 import { useContextStore } from '@/context';
 import { useAgent } from '@/hooks/agents';
-import { useIsDesktop } from '@/hooks/breakpoint';
 import { useConversation } from '@/hooks/conversation';
 import { useListTools, useShowUnauthedToolsModal } from '@/hooks/tools';
-import { useAgentsStore, useCitationsStore, useConversationStore, useParamsStore } from '@/stores';
+import { useCitationsStore, useConversationStore, useParamsStore } from '@/stores';
 import { OutputFiles } from '@/stores/slices/citationsSlice';
-import { cn, createStartEndKey, mapHistoryToMessages } from '@/utils';
+import { createStartEndKey, mapHistoryToMessages } from '@/utils';
 import { parsePythonInterpreterToolFields } from '@/utils/tools';
 
 const Chat: React.FC<{ agentId?: string; conversationId?: string }> = ({
@@ -30,10 +27,6 @@ const Chat: React.FC<{ agentId?: string; conversationId?: string }> = ({
   const { setConversation } = useConversationStore();
   const { addCitation, resetCitations, saveOutputFiles } = useCitationsStore();
   const { setParams, resetFileParams } = useParamsStore();
-  const {
-    agents: { isAgentsRightPanelOpen },
-  } = useAgentsStore();
-  const isDesktop = useIsDesktop();
 
   const { open, close } = useContextStore();
 
@@ -121,7 +114,7 @@ const Chat: React.FC<{ agentId?: string; conversationId?: string }> = ({
     saveOutputFiles(outputFilesMap);
   }, [conversation?.id, conversation?.messages.length, setConversation]);
 
-  const content = isLoading ? (
+  return isLoading ? (
     <div className="flex h-full flex-grow flex-col items-center justify-center">
       <Spinner />
     </div>
@@ -129,28 +122,6 @@ const Chat: React.FC<{ agentId?: string; conversationId?: string }> = ({
     <ConversationError error={error} />
   ) : (
     <Conversation conversationId={conversationId} agent={agent} tools={tools} startOptionsEnabled />
-  );
-
-  return (
-    <div className="flex w-full">
-      <div className="flex-grow">{content}</div>
-      <Transition
-        show={isAgentsRightPanelOpen || isDesktop}
-        as="div"
-        className={cn('dark:bg-volcanic-100', {
-          'mr-4 hidden w-[280px] flex-shrink-0 md:flex lg:w-[360px]': isDesktop,
-          'absolute inset-0': isAgentsRightPanelOpen || !isDesktop,
-        })}
-        enter="transition-all transform ease-in-out duration-300"
-        enterFrom="translate-x-full"
-        enterTo="translate-x-0"
-        leave="transition-all transform ease-in-out duration-300"
-        leaveFrom="translate-x-0 opacity-100"
-        leaveTo="translate-x-full opacity-0"
-      >
-        <AgentRightPanel />
-      </Transition>
-    </div>
   );
 };
