+++ src/interfaces/assistants_web/src/components/Conversation/index.tsx
@@ -136,7 +136,7 @@ const Conversation: React.FC<Props> = ({
 
   return (
     <div className="flex h-full w-full">
-      <div className="flex h-full w-full min-w-0 flex-col bg-marble-1000 dark:bg-volcanic-100">
+      <div className="flex h-full w-full min-w-0 flex-col rounded-lg bg-marble-1000 dark:bg-volcanic-100">
         <HotKeysProvider customHotKeys={chatHotKeys} />
         <Header agentId={agent?.id} />
