+++ src/interfaces/assistants_web/src/components/Layout.tsx
@@ -44,12 +44,7 @@ export const Layout: React.FC<Props> = ({ children }) => {
           <MobileHeader />
           <AgentLeftPanel className="hidden md:flex">{leftElement}</AgentLeftPanel>
           <section
-            className={cn(
-              'relative flex h-full min-w-0 flex-grow flex-col',
-              'rounded-lg border',
-              'border-marble-950 bg-marble-1000 dark:border-volcanic-100 dark:bg-volcanic-100',
-              'overflow-hidden'
-            )}
+            className={cn('relative flex h-full min-w-0 flex-grow flex-col', 'overflow-hidden')}
           >
             {mainElement}
           </section>
